+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Procedure when bumping the version number:
 - `euph.rooms.<name>.force_username` config option
 - `euph.rooms.<name>.password` config option
 - Key binding to change rooms sort order
+- Key bindings to connect to/disconnect from all rooms
+
+### Changed
+- Some key bindings in the rooms list
 
 ## v0.3.0 - 2022-08-22